Add base class initialization support and update first effects

+++ src/WS2812FX/effects/PopcornEffect.cpp
@@ -163,9 +163,9 @@ void PopcornEffect::renderKernel(double position, KernelData& kernel, WS2812FX* 
     uint8_t width = 1;
     uint8_t myWidth = (uint8_t)((double)(currentPos - prevPosition) / MM_PER_LED);
     if (currentPos > prevPosition) {
-        width = max(myWidth, 1);
+        width = max((uint8_t)myWidth, (uint8_t)1);
     } else if (prevPosition > currentPos) {
-        width = max(-myWidth, 1);
+        width = max((uint8_t)(-myWidth), (uint8_t)1);
     }
     
     // Render kernel using the strip's drawing function with current palette
